+++ packages/react-native-sdk/src/utils/StreamVideoRN/types.ts
@@ -110,10 +110,12 @@ export type StreamVideoConfig = {
      * }
      */
     createStreamVideoClient: () => Promise<StreamVideoClient | undefined>;
-    /** The callback that is called when a call is accepted, used for navigation */
-    navigateAcceptCall: () => void;
-    /** The callback that is called when a push notification is tapped but user did not press accept or decline, used for navigation */
-    navigateToIncomingCall: () => void;
+    /** @deprecated This method will be removed in the future. Please watch for incoming and outgoing calls in the root component of your app.
+        Please see https://getstream.io/video/docs/react-native/advanced/ringing-calls/#watch-for-incoming-and-outgoing-calls for more information */
+    navigateAcceptCall?: () => void;
+    /** @deprecated This method will be removed in the future. Please watch for incoming and outgoing calls in the root component of your app.
+        Please see https://getstream.io/video/docs/react-native/advanced/ringing-calls/#watch-for-incoming-and-outgoing-calls for more information */
+    navigateToIncomingCall?: () => void;
     /** Callback that is called when a non ringing push notification was tapped */
     onTapNonRingingCallNotification?: (
       call_cid: string,